+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.2.1 (2023-09-11)
+* Updated default vite version to 4
+* Added check for nested public/outDir
+* Added url prefix in dev manifest
+
 ## v0.2.0 (2023-09-11)
 * Simplified config generation
 * Removed plugins